+++ front/src/components/pc/templates/CategoryDetail/index.tsx
@@ -52,7 +52,7 @@ interface Props {
 }
 /**
  * テンプレート：カテゴリ一覧
- * デザイン途中なのでリファクタは後
+ * このページのみデザイン途中なのでリファクタはしない
  */
 const TemplatePCCategoryDetail: React.FC<Props> = (p: Props): JSX.Element => {
   const { color } = useColor();
