+++ front/src/modules/framework/createPages/categories.ts
@@ -1,0 +1,31 @@
+import path from 'path';
+import type { Actions } from 'gatsby';
+import { errorWrapper } from '../../common/error';
+import { ResCategory } from '../../interfaces/response/gatsbyNode/categories';
+
+/**
+ * ページ作成: カテゴリ一覧
+ */
+export const createPageCategories = (
+  categories: ResCategory[],
+  actions: Actions
+): void => {
+  try {
+    const { createPage } = actions;
+
+    const PAGE_PATH = `./src/pagesDynamic/categoryDetail.tsx`;
+    categories.forEach((c: ResCategory) => {
+      const id = c.id;
+
+      createPage({
+        path: `/category/${id}`,
+        component: path.resolve(PAGE_PATH),
+        context: {
+          id,
+        },
+      });
+    });
+  } catch (e) {
+    throw errorWrapper(e, 'ページ作成エラー:カテゴリ一覧');
+  }
+};
